+++ src/components/index.ts
@@ -5,6 +5,7 @@ export { default as Card } from './Card';
 export { default as InputWrapper } from './InputWrapper';
 export { default as Input } from './Input';
 export { default as Keypad } from './Keypad';
+export { default as CardSlider } from './CardSlider';
 export { default as CardSelector } from './CardSelector';
 export { default as PrevIcon } from './PrevIcon';
 export * from './Modal';
