+++ src/pages/Payment/components/TitleText/index.style.ts
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const TitleText = styled.div`
+  border-bottom: 2px solid ${({ theme }) => theme.color.grayBackground};
+  padding: 5px 0px;
+`;
