+++ src/pages/Payment/components/TitleText/index.tsx
@@ -1,0 +1,11 @@
+import * as S from './index.style';
+
+interface TitleTextProps {
+  label: string;
+}
+
+const TitleText = ({ label }: TitleTextProps) => {
+  return <S.TitleText>{label}</S.TitleText>;
+};
+
+export default TitleText;
