+++ src/pages/Payment/components/index.ts
@@ -1,3 +1,3 @@
 export { default as PayAmountInformation } from './PayAmountInformation';
-export { default as PayCardInformation } from './PayCardInformation';
+export { default as CardSelection } from './CardSelection';
 export { default as PayTermInformation } from './PayTermInformation';
