+++ src/pages/Payment/index.style.ts
@@ -42,8 +42,3 @@ export const TermInformationBottom = styled.div`
 
   font-size: 14px;
 `;
-
-export const TitleText = styled.div`
-  border-bottom: 2px solid ${({ theme }) => theme.color.grayBackground};
-  padding: 5px 0px;
-`;
